Refactor apps-service server structure
